- bugfixes - adjustment of spectra to mz-range of a profile
TODO:
- javadoc
- test with exactive data
- new cross validation

+++ src/Main/Main.java
@@ -24,6 +24,48 @@ public class Main {
 	
 	public static void main(String[] args) throws IOException, Exception {
 		
+//		String[] profilePath = readProfilePaths("/home/wens/cross-validation/cross-validation-1");
+//		
+//		SpectraMatrix data = Reader.readData(profilePath, 
+//				"/home/wens/cross-validation/cross-validation-1", 
+//				2.0, 
+//				"Mini 11");
+//		PCADataSet pca_data = PCA.performPCA(data, 0.9);
+//		LDADataSet lda_data = LDA.performLDA(pca_data, data);
+//		
+//		// create profile
+//		ProfileBuilder.build(
+//				pca_data, 
+//				lda_data,
+//				data, 
+//				"Mini 11", 
+//				"/home/wens/cross-validation/cross-validation-1", 
+//				"/home/wens/xyz.profile", 
+//				1.0);
+		
+		// open profile
+//		Profile profile = Reader.readProfile("/home/wens/xyz.profile");
+		
+//		Spectrum spectrum = new Spectrum(
+//				"/home/wens/cross-validation/cross-validation-2-destination/data/Dakapo_Accent_54AVG5.csv", 
+//				null, 
+//				(int)profile.getBinSize(), 
+//				"Mini 11");
+//		ClassificationResult res_ed = profile.euclideanDistance(spectrum);
+//		
+//		// open folder with csv files to classify
+//		String[] csv = Reader.readFolder("/home/wens/cross-validation/cross-validation-2-destination/data");
+//		
+//		for(int i=0; i<csv.length; i++){
+//			System.out.println(csv[i]);
+//			Spectrum spectrum = new Spectrum(csv[i], null, (int)profile.getBinSize(), "Mini 11");
+//			ClassificationResult res_ed = profile.euclideanDistance(spectrum);
+//			spectrum = new Spectrum(csv[i], null, (int)profile.getBinSize(), "Mini 11");
+//			ClassificationResult res_md = profile.mahalanobisDistance(spectrum);
+//			spectrum = new Spectrum(csv[i], null, (int)profile.getBinSize(), "Mini 11");
+//			ClassificationResult res_lda = profile.ldaCoefficient(spectrum);
+//		}
+		
 		/*
 		args[0] = root directory path
 		args[1] = binsize
@@ -69,11 +111,11 @@ public class Main {
 		
 		// classify
 		for(int i=0; i<csv.length; i++){
-			Spectrum spectrum = new Spectrum(csv[i], null, (int)profile.getBinSize(), "Mini 11");
+			Spectrum spectrum = new Spectrum(csv[i], null, profile.getBinSize(), "Mini 11");
 			ClassificationResult res_ed = profile.euclideanDistance(spectrum);
-			spectrum = new Spectrum(csv[i], null, (int)profile.getBinSize(), "Mini 11");
+			spectrum = new Spectrum(csv[i], null, profile.getBinSize(), "Mini 11");
 			ClassificationResult res_md = profile.mahalanobisDistance(spectrum);
-			spectrum = new Spectrum(csv[i], null, (int)profile.getBinSize(), "Mini 11");
+			spectrum = new Spectrum(csv[i], null, profile.getBinSize(), "Mini 11");
 			ClassificationResult res_lda = profile.ldaCoefficient(spectrum);
 			writer.println(spectrum.getFilename() + "\t" 
 					+ res_ed.getAssignedClass() + "\t"
